pyproject.toml
[project]
name = "desafio-api"
version = "0.1.0"
description = ""
authors = [
    {name = "MihaelOli",email = "michaeloliveira38@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "bcrypt (==4.0.1)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
